Don't force compile everything as Objective-C++ on macOS
