+++ Source/Core/Common/CMakeLists.txt
@@ -82,7 +82,7 @@ endif()
 if(WIN32)
     set(SRCS ${SRCS} GL/GLInterface/WGL.cpp)
 elseif(APPLE)
-    set(SRCS ${SRCS} GL/GLInterface/AGL.cpp)
+    set(SRCS ${SRCS} GL/GLInterface/AGL.mm)
 elseif(USE_X11)
     if (NOT USE_EGL)
         set(SRCS ${SRCS} GL/GLInterface/GLX.cpp)
