+++ Source/Core/Core/HW/WiimoteReal/IOdarwin.mm
@@ -4,6 +4,7 @@
 #include "Common/Common.h"
 #include "Common/Logging/Log.h"
 #include "Core/HW/WiimoteEmu/WiimoteHid.h"
+#include "Core/HW/WiimoteReal/IOdarwin_private.h"
 
 @interface SearchBT : NSObject
 {
